Relatve paths to CSS assets

+++ src/script/3ify.py
@@ -7,65 +7,111 @@ try:
 except:
    import simplejson
 
-import os, codecs, shutil, copy, re
+import os, codecs, shutil, copy, re, sys
 
 class Threeify(object):
 
     def __init__(self):
+        l = len(sys.argv)
 
-        VERSION         = '2.8.0'
-        TNT_VERSION     = '2'
+        if l > 1:
+            VERSION     = sys.argv[1]
+        else:
+            VERSION     = '2.8.1'
 
-        SRC_DIR        = '../'
+        if l > 2:
+            YUI3_VERSION     = sys.argv[2]
+        else:
+            YUI3_VERSION     = '3.2.0'
+
+        if l > 3:
+            TNT_VERSION     = sys.argv[3]
+        else:
+            TNT_VERSION     = '3'
+
+        SRC_DIR         = './'
         SRC_FILE        = VERSION + '_expanded.json'
 
         # YUI2_DIR        = '../../yui2/build/'
         YUI2_DIR        = 'lib/' + VERSION + '/build/'
+        LOCAL_CSS_DIR   = 'lib/localcss/' + VERSION + '/build/'
 
         TEMPLATE_DIR    = 'template'
         TEMPLATE_FILE   = 'yui3mod.js'
         SUPERSEDED_FILE = 'superseded.js'
 
-        DEST_DIR        = 'build_tmp'
+        DEST_DIR        = 'build_tmp.cdn'
         dest_path      = os.path.abspath(DEST_DIR)
 
-        if not os.path.exists(dest_path):         
+        GIT_DIR        = 'build_tmp.local'
+        git_path       = os.path.abspath(GIT_DIR)
+
+        if not os.path.exists(dest_path):
             os.mkdir(dest_path)
+
+        if not os.path.exists(git_path):
+            os.mkdir(git_path)
 
         DEST_JSON       = VERSION + '.json'
 
-        # BUILD_DIR       = 'build_tmp/' + VERSION + '_build/'
-        # build_path     = os.path.abspath(BUILD_DIR)
+        build_path       = os.path.join(dest_path, '2in3.' + TNT_VERSION)
+        git_build_path   = os.path.join(git_path, '2in3.' + TNT_VERSION)
 
-        build_path       = os.path.join(dest_path, TNT_VERSION)
 
-        if not os.path.exists(build_path):         
+        if not os.path.exists(build_path):
             os.mkdir(build_path)
+
+        if not os.path.exists(git_build_path):
+            os.mkdir(git_build_path)
 
         build_path       = os.path.join(build_path, VERSION)
+        git_build_path   = os.path.join(git_build_path, VERSION)
 
-        if not os.path.exists(build_path):         
-            os.mkdir(build_path)
+        if os.path.exists(build_path):
+            shutil.rmtree(build_path, True)
+
+        if os.path.exists(git_build_path):
+            shutil.rmtree(git_build_path, True)
+
+        os.mkdir(build_path)
+        os.mkdir(git_build_path)
+
+        build_path       = os.path.join(build_path, 'build')
+        git_build_path   = os.path.join(git_build_path, 'build')
+
+        os.mkdir(build_path)
+        os.mkdir(git_build_path)
 
         TOKENS = {}
 
-        TOKENS['name'] =    '{ /* NAME */ }' 
-        TOKENS['superseded'] = '{ /* SUPERSEDED */ }' 
-        TOKENS['content'] = '{ /* SOURCE */ }' 
-        TOKENS['data'] =    '{ /* DATA */ }' 
-        TOKENS['version'] =    '{ /* VERSION */ }' 
+        TOKENS['name']       = '{ /* NAME */ }'
+        TOKENS['superseded'] = '{ /* SUPERSEDED */ }'
+        TOKENS['content']    = '{ /* SOURCE */ }'
+        TOKENS['data']       = '{ /* DATA */ }'
+        TOKENS['module']     = '{ /* MODULE */ }'
+        TOKENS['version']    = '{ /* VERSION */ }'
 
-        src_path       = os.path.abspath(SRC_DIR)
+        src_path        = os.path.abspath(SRC_DIR)
         print src_path
-        yui2_path      = os.path.abspath(YUI2_DIR)
 
-        template_path  = os.path.abspath(TEMPLATE_DIR)
+        yui2_path       = os.path.abspath(YUI2_DIR)
+        local_css_path  = os.path.abspath(LOCAL_CSS_DIR)
 
+        assets_path     = os.path.join(yui2_path, 'assets')
+        assets_dest     = os.path.join(build_path, 'assets')
+        git_assets_dest = os.path.join(git_build_path, 'assets')
+
+        template_path   = os.path.abspath(TEMPLATE_DIR)
 
         print dest_path
 
-        shutil.rmtree(build_path, True)
-        shutil.copytree(yui2_path, build_path)
+        if os.path.exists(assets_path):
+            shutil.copytree(assets_path, assets_dest)
+            shutil.copytree(assets_path, git_assets_dest)
+
+
+        # shutil.copytree(yui2_path, build_path)
+        # shutil.copytree(yui2_path, git_build_path)
 
         def readFile(path, file):
             return codecs.open(os.path.join(path, file), "r", "utf-8").read()
@@ -79,6 +125,10 @@ class Threeify(object):
 
         # write the raw module json
         out = codecs.open(os.path.join(dest_path, DEST_JSON), 'w', 'utf-8')
+        out.writelines(jsonstr)
+        out.close()
+
+        out = codecs.open(os.path.join(git_path, DEST_JSON), 'w', 'utf-8')
         out.writelines(jsonstr)
         out.close()
 
@@ -113,6 +163,13 @@ class Threeify(object):
             except:
                 print 'WARNING: could not find file ' + path_min
                 continue;
+
+            if not type == 'js':
+                try:
+                    local_css_content  = readFile(local_css_path, path_min)
+                except:
+                    print 'WARNING: could not find file ' + path_min
+                    continue;
 
             modcopy = copy.deepcopy(mod)
 
@@ -156,6 +213,8 @@ class Threeify(object):
                         supresult = superseded
                         supresult = supresult.replace(TOKENS['name'], supname)
                         supresult = supresult.replace(TOKENS['data'], supdata)
+                        supresult = supresult.replace(TOKENS['module'], name)
+                        supresult = supresult.replace(TOKENS['version'], YUI3_VERSION)
                         supresult = supresult.replace(TOKENS['content'], '')
                         supcontent += supresult
 
@@ -170,16 +229,25 @@ class Threeify(object):
                     result = content
 
 
-                out = codecs.open(os.path.join(build_path, path), 'w', 'utf-8')
-                out.truncate(0)
-                out.writelines(result)
-                out.close()
+                # out = codecs.open(os.path.join(build_path, path), 'w', 'utf-8')
+                # out.truncate(0)
+                # out.writelines(result)
+                # out.close()
+
+                # out = codecs.open(os.path.join(git_build_path, path), 'w', 'utf-8')
+                # out.truncate(0)
+                # out.writelines(result)
+                # out.close()
 
                 # default path
 
                 default_path = os.path.join(build_path, name)
-                if not os.path.exists(default_path):         
+                if not os.path.exists(default_path):
                     os.mkdir(default_path)
+
+                git_default_path = os.path.join(git_build_path, name)
+                if not os.path.exists(git_default_path):
+                    os.mkdir(git_default_path)
 
                 if name.count('skin-') > 0:
                     pkg = re.search('^[^\/]*', path)
@@ -187,10 +255,11 @@ class Threeify(object):
                     original_path = os.path.join(build_path, pkg.group(0))
                     asset_path = os.path.join(original_path, 'assets')
                     print "asset path: " + asset_path
-                    if os.path.exists(asset_path):         
+                    if os.path.exists(asset_path):
                         print "copying to: " + default_path
                         # try:
                         shutil.copytree(asset_path, os.path.join(default_path, 'assets'), True)
+                        shutil.copytree(asset_path, os.path.join(git_default_path, 'assets'), True)
                         # except:
                             # print 'skipping duplicate dir copy'
 
@@ -199,25 +268,52 @@ class Threeify(object):
                     out.truncate(0)
                     out.writelines(result)
                     out.close()
-                    out = codecs.open(os.path.join(default_path, name + '-debug.' + type), 'w', 'utf-8')
-                    out.truncate(0)
-                    out.writelines(result)
-                    out.close()
-                    out = codecs.open(os.path.join(default_path, name + '.' + type), 'w', 'utf-8')
+
+                    out = codecs.open(os.path.join(git_default_path, name + filter + type), 'w', 'utf-8')
                     out.truncate(0)
                     out.writelines(result)
                     out.close()
 
+                    out = codecs.open(os.path.join(default_path, name + '-debug.' + type), 'w', 'utf-8')
+                    out.truncate(0)
+                    out.writelines(result)
+                    out.close()
+
+                    out = codecs.open(os.path.join(git_default_path, name + '-debug.' + type), 'w', 'utf-8')
+                    out.truncate(0)
+                    out.writelines(local_css_content)
+                    out.close()
+
+                    out = codecs.open(os.path.join(default_path, name + '.' + type), 'w', 'utf-8')
+                    out.truncate(0)
+                    out.writelines(local_css_content)
+                    out.close()
+
+                    out = codecs.open(os.path.join(git_default_path, name + '.' + type), 'w', 'utf-8')
+                    out.truncate(0)
+                    out.writelines(local_css_content)
+                    out.close()
+
                     # we need the full directory structure to get the relative path to the skin file
                     default_path = os.path.join(default_path, 'assets')
-                    if not os.path.exists(default_path):         
+                    if not os.path.exists(default_path):
                         os.mkdir(default_path)
                     default_path = os.path.join(default_path, 'skins')
-                    if not os.path.exists(default_path):         
+                    if not os.path.exists(default_path):
                         os.mkdir(default_path)
                     default_path = os.path.join(default_path, 'sam')
-                    if not os.path.exists(default_path):         
+                    if not os.path.exists(default_path):
                         os.mkdir(default_path)
+
+                    git_default_path = os.path.join(git_default_path, 'assets')
+                    if not os.path.exists(git_default_path):
+                        os.mkdir(git_default_path)
+                    git_default_path = os.path.join(git_default_path, 'skins')
+                    if not os.path.exists(git_default_path):
+                        os.mkdir(git_default_path)
+                    git_default_path = os.path.join(git_default_path, 'sam')
+                    if not os.path.exists(git_default_path):
+                        os.mkdir(git_default_path)
 
                 # copy skin assets
 
@@ -226,14 +322,40 @@ class Threeify(object):
                 out.writelines(result)
                 out.close()
 
+                out = codecs.open(os.path.join(git_default_path, name + filter + type), 'w', 'utf-8')
+                out.truncate(0)
+                if type == 'js':
+                    out.writelines(result)
+                else:
+                    out.writelines(local_css_content)
+
+                out.close()
+
                 if len(paths) < 3:
                     out = codecs.open(os.path.join(default_path, name + '-debug.' + type), 'w', 'utf-8')
                     out.truncate(0)
                     out.writelines(result)
                     out.close()
+
+                    out = codecs.open(os.path.join(git_default_path, name + '-debug.' + type), 'w', 'utf-8')
+                    out.truncate(0)
+                    if type == 'js':
+                        out.writelines(result)
+                    else:
+                        out.writelines(local_css_content)
+                    out.close()
+
                     out = codecs.open(os.path.join(default_path, name + '.' + type), 'w', 'utf-8')
                     out.truncate(0)
                     out.writelines(result)
+                    out.close()
+
+                    out = codecs.open(os.path.join(git_default_path, name + '.' + type), 'w', 'utf-8')
+                    out.truncate(0)
+                    if type == 'js':
+                        out.writelines(result)
+                    else:
+                        out.writelines(local_css_content)
                     out.close()
 
         # for name, mod in src.iteritems():
